turned the 3d array rubbish into flat vec, broke parts in process

+++ src/settings.rs
@@ -1,4 +1,4 @@
-pub const TRANSLATIONS_VON: [[i8; 3]; 6] = [
+pub const TRANSLATIONS_VON: [[i16; 3]; 6] = [
     [1, 0, 0],
     [-1, 0, 0],
     [0, 1, 0],
@@ -8,7 +8,7 @@ pub const TRANSLATIONS_VON: [[i8; 3]; 6] = [
 ];
 
 
-pub const TRANSLATIONS_MOORE: [[i8 ;3];26] = [
+pub const TRANSLATIONS_MOORE: [[i16 ;3];26] = [
     [1, 0, 0],
     [1, 1, 0],
     [1, -1, 0],
